requirements.txt
requests==2.31.0
pandas==2.1.4
streamlit==1.28.2
tushare==1.4.24
python-dotenv==1.0.0
regex==2023.10.3
plotly==5.18.0
openai==1.12.0
httpx==0.27.2
